- AppRootViewController + InitialSetupViewController: move initial setup view into own view controller - InitialSetupViewController: implement status bar style support for consistent status bar colors also pre-iOS 17
